Add files via upload
Frontend: Using HTML and CSS, I created a clean and elegant interface for the form, with animations and modern design.

Backend: Developed in Python, using the qrcode or pyzbar libraries, to generate the QR Code from the link or text provided by the user.

Functionality: The user enters a link or text in the input field, and the system automatically generates a QR Code that can be downloaded. The solution is simple and efficient for generating customized and practical QR Codes.